requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
flake8>=3.7.9,<3.8.0
psycopg2>=2.7.5,<2.8.0
django-cors-headers>=3.2.1,<3.3.0
